refactor(ProductSelector): remove old implementation
